+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -210,6 +210,8 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   private cursorOnLeaflet = false;
   /* Circle drawn when trackingposition to show accuracy of coordinate */
   private proximityCircle;
+  /* Circle drawn when trackingposition to show coordinate */
+  private proximityCenter;
   /* Timeout to reset clearwatch*/
   private clearWatchTimeoutId;
 
@@ -858,10 +860,20 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
    * Adds proximity circle to show the accuracy of received coordinates
    */
   private showProximityCircle(coords: number[], accuracy: number): void {
+    let zoomLevel = 15;
     this.removeProximityCircle();
-    this.proximityCircle = this.locationViewerService.L.circle(coords, accuracy, { opacity: 0 });
+    this.proximityCircle = this.locationViewerService.L.circle(coords, accuracy, { opacity: 1, color: '#FFFFFF', fillColor: '#3388ff', className: 'proximity' });
+    this.proximityCenter = this.locationViewerService.L.circle(coords, 15, { fillOpacity: 1, color: '#FFFFFF', fillColor: '#0057b7', className: 'proximity-center' });
     this.proximityCircle.addTo(this.leafletMap.map);
-    this.leafletMap.setView(coords, 15);
+    this.proximityCenter.addTo(this.leafletMap.map);
+
+    const boundZoom = this.leafletMap.map.getBoundsZoom(this.proximityCircle.getBounds());
+    if (boundZoom < zoomLevel)
+    {
+      zoomLevel = boundZoom;
+    }
+    
+    this.leafletMap.setView(coords, zoomLevel);
   }
 
   /**
@@ -871,6 +883,11 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
     if (this.proximityCircle) {
       this.proximityCircle.remove();
       this.proximityCircle = null;
+    }
+
+    if (this.proximityCenter) {
+      this.proximityCenter.remove();
+      this.proximityCenter = null;
     }
   }